added deterministic ODE simulator to Dizzy
git-svn-id: file:///Users/sramsey/Desktop/DizzyRepo/ISBJava_Dizzy@1871 5021c4a6-a05a-4b22-8634-7b27b77db61f

+++ ISBJava_Dizzy/trunk/bolouri/java/isb/chem/sbw/MarkupLanguageModelInstanceExporter.java
@@ -185,7 +185,7 @@ public class MarkupLanguageModelInstanceExporter implements IModelInstanceExport
                 Compartment compartment = species.getCompartmentCopy();
                 String compartmentName = compartment.getName();
                 speciesElement.setAttribute(ATTRIBUTE_COMPARTMENT, compartmentName);
-                long initialSpeciesPopulation = 0;
+                double initialSpeciesPopulation = 0.0;
                 try
                 {
                     initialSpeciesPopulation = pInitialSpeciesPopulations.getSpeciesPopulation(speciesName);
